+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 * When Order template includes hidden fields they are also hidden from the view pane. Refs UIOR-856.
 * Order - Allow user to choose what columns display for Order lines. Refs UIOR-878.
 * Add estimated price to POL grid view on order record. Refs UIOR-868.
+* Check for potential duplicate order lines. Refs UIOR-864.
 
 ## [3.0.3](https://github.com/folio-org/ui-orders/tree/v3.0.3) (2021-12-08)
 [Full Changelog](https://github.com/folio-org/ui-orders/compare/v3.0.2...v3.0.3)